removed extention from url, added 404 and htaccess. contact form needs fixing :)
