--- 404.html
@@ -1,1 +1,0 @@
-<h1>404 Error</h1>
