+++ home.php
@@ -1,0 +1,201 @@
+<!doctype html>
+<html class="no-js" lang="en">
+
+<head>
+  <meta charset="utf-8">
+  <meta http-equiv="x-ua-compatible" content="ie=edge">
+  <meta name="viewport" content="width=device-width,initial-scale=1">
+
+  <title>Awake Birmingham</title>
+  <meta name="description" content="We support parents, students and schools to resolve issues that negatively impact on the students' education including: Fixed Term or Permanent Exclusions, Managed Moves, Pastoral Support, Reviews, School Meetings and Special Educational Needs.
+">
+  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+
+  <!-- Icons -->
+  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.6.3/css/all.css" integrity="sha384-UHRtZLI+pbxtHCWp1t77Bi1L4ZtiqrqD80Kn4Z8NTSRyMA2Fd33n5dQ8lWUE00s/"
+    crossorigin="anonymous">
+
+  <!-- Place favicon.ico in the root directory -->
+  <link rel="manifest" href="manifest.json">
+  <link rel="apple-touch-icon" href="/img/icons/icon.png">
+  <!-- Place favicon.ico in the root directory -->
+
+  <!--Bootstrap -->
+  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.2.1/css/bootstrap.min.css" integrity="sha384-GJzZqFGwb1QTTN6wy59ffF1BuGJpLSa9DkKMp0DgiMDm4iYMj70gZWKYbI706tWS"
+    crossorigin="anonymous">
+
+  <link rel="stylesheet" href="/sass/main.css">
+</head>
+
+<body>
+  <header>
+    <nav id="mobnav" class="navbar navbar-expand-lg navbar-light bg-light">
+      <a class="navbar-brand" href="/home"><img src="/img/logo.png" alt="Awake Birmingham logo">Awake Birmingham</a>
+      <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavAltMarkup"
+        aria-controls="navbarNavAltMarkup" aria-expanded="false" aria-label="Toggle navigation">
+        <span class="navbar-toggler-icon"></span>
+      </button>
+      <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
+        <div class="navbar-nav">
+          <a class="nav-item nav-link active" href="/home">Home <span class="sr-only">(current)</span></a>
+          <a class="nav-item nav-link" href="/who-we-are">Who We Are</a>
+          <a class="nav-item nav-link" href="/what-we-do">What We Do</a>
+          <a class="nav-item nav-link" href="/contact-us/contact-us">Contact Us</a>
+        </div>
+      </div>
+    </nav>
+
+  </header>
+
+  <main>
+    <section id="intro">
+      <div id="intro-overlay">
+        <h1 id="mobintro">We support parents, students and schools to resolve issues that negatively impact on the
+          students' education.</h1>
+      </div>
+      <h1 id="deskintro">We support parents, students and schools to resolve issues that negatively impact on the
+        students' education.</h1>
+    </section>
+
+    <!-- TESTIMONIALS -->
+    <section id="testimonials">
+      <div id="carouselExampleIndicators" class="carousel slide" data-ride="carousel">
+        <ol class="carousel-indicators">
+          <li data-target="#carouselExampleIndicators" data-slide-to="0" class="active"></li>
+          <li data-target="#carouselExampleIndicators" data-slide-to="1"></li>
+          <li data-target="#carouselExampleIndicators" data-slide-to="2"></li>
+          <li data-target="#carouselExampleIndicators" data-slide-to="3"></li>
+          <li data-target="#carouselExampleIndicators" data-slide-to="4"></li>
+          <li data-target="#carouselExampleIndicators" data-slide-to="5"></li>
+          <li data-target="#carouselExampleIndicators" data-slide-to="6"></li>
+          <li data-target="#carouselExampleIndicators" data-slide-to="7"></li>
+          <li data-target="#carouselExampleIndicators" data-slide-to="8"></li>
+          <li data-target="#carouselExampleIndicators" data-slide-to="9"></li>
+        </ol>
+        <div class="carousel-inner">
+          <div class="carousel-item active">
+            <p>We have had the toughest time but we have hung in there together. We have laughed and we have cried.
+              He nearly lost out on three very valuable school experiences, but he did get to go on a trip, he got
+              brilliant SAT results and he got to go to prom. No matter how dark your days get remember if you make
+              it through the night the sun will always rise again in the morning. Thank you Awake.</p>
+            <small>Myrah</small>
+          </div>
+          <div class="carousel-item">
+            <p>Thank you Awake for giving me that confidence. When I came to see you I was a bit wobbly.</p>
+            <br>
+            <small>Anon</small>
+          </div>
+          <div class="carousel-item">
+            <p>Tremendous respect for the support you gave. So happy to report how amazingly different his education
+              is. It’s like having a new child who has an amazing relationship with his teachers and continues to
+              thrive and develop. I can only
+              advise parents to not fear change a fresh start really does overcome educational obstacles.</p>
+            <br>
+            <small>Tina</small>
+          </div>
+          <div class="carousel-item">
+            <p>I had an amazing experience with the advice and support of Awake Bham... permanent exclusion over
+              turned. Keep up the amazing work as our children and parents need a voice and with yourselves that is
+              what we have. THANK YOU ALL for everything.</p>
+            <br>
+            <small>Anon</small>
+          </div>
+          <div class="carousel-item">
+            <p>Thank God there's a place of support to give parents and children a voice.</p>
+            <br>
+            <small>Anon</small>
+          </div>
+          <div class="carousel-item">
+            <p>Really appreciate advice given from Awake Bham. You have helped my stress levels tremendously and
+              given me a better understanding of what can be done to help my son's education.</p>
+            <br>
+            <small>Anon</small>
+          </div>
+          <div class="carousel-item">
+            <p>Today I received support from this fantastic organisation. I can honestly say in my opinion it can be
+              very
+              daunting going through authorities procedures especially if you don't have the relevant training. I felt
+              alone and scared almost like I was backed into a corner but with the support of Awake things seem a
+              lot less stressful. I want to say a big massive thank you for your help and I would recommend this
+              service
+              to anyone. Thank you.</p>
+            <br>
+            <small>Anon</small>
+          </div>
+          <div class="carousel-item">
+            <p>There comes a time in your life when you need help, not just someone saying 'I can help' but an
+              understanding person with the specific set of tools for that particular job. Not only does Jalika have
+              the historical wisdom for the task on hand but she executes her expertise with purpose,
+              courtesy and professionalism. I can whole-heartedly recommend Awake Bham as a community organisation that
+              puts their clients not only first but at the heart of their interventions. I am therefore truly grateful
+              and honoured to have them in my corner when I need help, support and understanding the most. Much respect
+              and long may you continue to make changes within our communities. Peace.</p>
+            <br>
+            <small>Anon</small>
+          </div>
+          <div class="carousel-item">
+            <p>Awake Bham thank you so much for sharing your expertise with others for better outcomes for young people
+              and families when it comes to education. A fine example of dedicated individuals who come together to
+              support the community. #DoingTheDo</p>
+            <br>
+            <small>Anon</small>
+          </div>
+          <div class="carousel-item">
+            <p>Thank you so much for your help yesterday, your presence not only gave me strength but also made a huge
+              impact. I hope this teacher continues to acknowledge my son's autism traits as
+              this will help my son tremendously. I wasn't sure if I got off track but thanks to you he let me explain
+              about my son's routine and behaviour and now he does have some insight. I greatly appreciate your help</p>
+            <br>
+            <small>Anon</small>
+          </div>
+        </div>
+        <a class="carousel-control-prev" href="#carouselExampleIndicators" role="button" data-slide="prev">
+          <span class="carousel-control-prev-icon" aria-hidden="true"></span>
+          <span class="sr-only">Previous</span>
+        </a>
+        <a class="carousel-control-next" href="#carouselExampleIndicators" role="button" data-slide="next">
+          <span class="carousel-control-next-icon" aria-hidden="true"></span>
+          <span class="sr-only">Next</span>
+        </a>
+      </div>
+    </section>
+  </main>
+
+  <footer class="row">
+    <div class="col-12 col-md-12 col-lg-12">
+      <a href="mailto:infoawake16@gmail.com">infoawake16@gmail.com</a>
+    </div>
+    <div class="col-12 col-md-4 col-lg-4 contact-links jalika">
+      Jalika: <a href="tel:+447538606045">(+44) 7538 606045</a>
+    </div>
+    <div class="col-12 col-md-4 col-lg-4 contact-links">
+      Imani: <a href="tel:+447828778484">(+44) 7828 778484</a>
+    </div>
+    <div class="col-12 col-md-4 col-lg-4 contact-links shennice">
+      Shennice: <a href="tel:+447999731982">(+44) 7999 731982</a>
+    </div>
+    <div class="col-12">
+      <a href="https://www.facebook.com/awa.ke.31" target="_blank"><i class="fab fa-facebook-square"></i></a>
+
+      <a href="https://www.instagram.com/bhamawake/" target="_blank"><i class="fab fa-instagram"></i></a>
+
+      <a href="https://www.linkedin.com/in/awake-birmingham-5a9849174" target="_blank"><i class="fab fa-linkedin"></i></a>
+    </div>
+    <div class="col-12">
+      <p>&copy;
+        <?php echo date("Y");?> Awake</p>
+    </div>
+  </footer>
+
+
+
+  <!--Bootstrap JS-->
+  <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
+    crossorigin="anonymous"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.6/umd/popper.min.js" integrity="sha384-wHAiFfRlMFy6i5SRaxvfOCifBUQy1xHdJ/yoi7FRNXMRBu5WHdZYu1hA6ZOblgut"
+    crossorigin="anonymous"></script>
+  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.2.1/js/bootstrap.min.js" integrity="sha384-B0UglyR+jN6CkvvICOB2joaf5I4l3gm9GU6Hc1og6Ls7i6U/mkkaduKaBhlAXv9k"
+    crossorigin="anonymous"></script>
+</body>
+
+</html>
