+++ htaccess.txt
@@ -1,0 +1,23 @@
+ErrorDocument 404 /404.php
+
+# Apache Rewrite Rules
+ <IfModule mod_rewrite.c>
+  Options +FollowSymLinks
+  RewriteEngine On
+  RewriteBase /
+
+  RewriteRule ^contact.php - [PT]
+  
+# Add trailing slash to url
+  RewriteCond %{REQUEST_FILENAME} !-f
+  RewriteCond %{REQUEST_URI} !(\.[a-zA-Z0-9]{1,5}|/|#(.*))$
+  RewriteRule ^(.*)$ $1/ [R=301,L]
+
+# Remove .php-extension from url
+  RewriteCond %{REQUEST_METHOD} !POST [NC]
+  RewriteCond %{REQUEST_FILENAME} !-d
+  RewriteCond %{REQUEST_FILENAME}\.php -f
+  RewriteRule ^([^\.]+)/$ $1.php 
+
+# End of Apache Rewrite Rules
+ </IfModule>
