+++ manifest.json
@@ -1,0 +1,52 @@
+{
+    "short_name": "Awake",
+    "name": "Awake Birmingham",
+    "description": "Awake supports parents, students and schools to resolve issues that negatively impact on the students' education.",
+    "icons": [
+        {
+          "src": "images/icons/icon-72x72.png",
+          "sizes": "72x72",
+          "type": "image/png"
+        },
+        {
+          "src": "img/icons/icon-96x96.png",
+          "sizes": "96x96",
+          "type": "image/png"
+        },
+        {
+          "src": "img/icons/icon-128x128.png",
+          "sizes": "128x128",
+          "type": "image/png"
+        },
+        {
+          "src": "img/icons/icon-144x144.png",
+          "sizes": "144x144",
+          "type": "image/png"
+        },
+        {
+          "src": "img/icons/icon-152x152.png",
+          "sizes": "152x152",
+          "type": "image/png"
+        },
+        {
+          "src": "img/icons/icon-192x192.png",
+          "sizes": "192x192",
+          "type": "image/png"
+        },
+        {
+          "src": "img/icons/icon-384x384.png",
+          "sizes": "384x384",
+          "type": "image/png"
+        },
+        {
+          "src": "img/icons/icon-512x512.png",
+          "sizes": "512x512",
+          "type": "image/png"
+        }
+      ],
+    "start_url": "/home",
+    "background_color": "#AA076B",
+    "display": "standalone",
+    "scope": "/",
+    "theme_color": "#AA076B"
+  }
