+++ sass/404.css
@@ -1,0 +1,2 @@
+@keyframes fadein{from{opacity:0}to{opacity:1}}@-moz-keyframes fadein{from{opacity:0}to{opacity:1}}@-webkit-keyframes fadein{from{opacity:0}to{opacity:1}}@-o-keyframes fadein{from{opacity:0}to{opacity:1}}body main .button-container{margin:30px auto;width:80%}body main .button-container #cta-button,body main .button-container #send-button,body main .button-container #four-button{background-color:#DE0D92;border-color:#DE0D92}body main .button-container #cta-button:hover,body main .button-container #send-button:hover,body main .button-container #four-button:hover{background-color:#E439A5;border-color:#E439A5}body main .form-button #cta-button,body main .form-button #send-button,body main .form-button #four-button{background-color:#DE0D92;border-color:#DE0D92}body main .form-button #cta-button:hover,body main .form-button #send-button:hover,body main .form-button #four-button:hover{background-color:#E439A5;border-color:#E439A5}body main .form-button #send-button{width:100%;margin:0 auto}body main .four-button-container{margin:50px auto 70px auto}body main .four-button-container #cta-button,body main .four-button-container #send-button,body main .four-button-container #four-button{background-color:#DE0D92;border-color:#DE0D92}body main .four-button-container #cta-button:hover,body main .four-button-container #send-button:hover,body main .four-button-container #four-button:hover{background-color:#E439A5;border-color:#E439A5}body main .four-button-container #four-button{width:25% !important;margin:0 auto}@media only screen and (min-width: 2501px) and (max-width: 3000px){body main .form-button #cta-button,body main .form-button #send-button,body main .form-button #four-button{background-color:#DE0D92;border-color:#DE0D92}body main .form-button #cta-button:hover,body main .form-button #send-button:hover,body main .form-button #four-button:hover{background-color:#E439A5;border-color:#E439A5}body main .form-button #send-button{width:100%;margin-top:15px}}html{width:100%;height:100%;margin:0 auto}html body{max-width:100%;min-height:100%;margin:0 auto;overflow:hidden;background:#AA076B;background:-webkit-linear-gradient(to right, #61045F, #AA076B);background:linear-gradient(to right, #61045F, #AA076B);color:#D1D1D1}html body #wrapper{width:100%;height:100vh;display:flex;flex-flow:column nowrap;justify-content:center;align-items:center}html body #wrapper #content{text-align:center}html body #wrapper #content h1{font-size:5rem;margin:20px 0 20px 0;font-weight:bold}html body #wrapper #content h2{font-weight:700;font-size:2rem;margin:30px 0;color:white}html body #wrapper #content #socialmedia{margin:30px auto}html body #wrapper #content #socialmedia .fab{font-size:1.8rem;margin:0 10px;color:#DE0D92}html body #wrapper #content #socialmedia .fab:visited{color:#DE0D92}html body #wrapper #content #socialmedia .fab:hover{color:#330d22}html body #wrapper #content #socialmedia .fab:active{color:#330d22}
+/*# sourceMappingURL=404.css.map */
