+++ sass/404.scss
@@ -1,0 +1,6 @@
+@import "base/variables";
+@import "base/mixins";
+
+@import "component/buttons";
+
+@import "layout/404-main";
