+++ sass/base/_variables.scss
@@ -19,3 +19,9 @@ $links-icon-fh: #DE0D92;
 /* BUTTON COLORS */
 $button-primary: #DE0D92;
 $button-hover: #E439A5;
+
+/* 404 PAGE */
+$background-gradient: #61045F;
+$four-text-color: #D1D1D1;
+$four-links-icon-f: #DE0D92;
+$four-links-icon-fh: #330d22;
