+++ sass/layout/_404-main.scss
@@ -1,0 +1,52 @@
+html {
+    width: 100%;
+    height: 100%;
+    margin: 0 auto;
+
+    body {
+        max-width: 100%;
+        min-height: 100%;
+        margin: 0 auto;
+        overflow: hidden;
+        background: $primary-color;
+        /* fallback for old browsers */
+        background: -webkit-linear-gradient(to right, $background-gradient, $primary-color);
+        /* Chrome 10-25, Safari 5.1-6 */
+        background: linear-gradient(to right, $background-gradient, $primary-color);
+        /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
+        color: $four-text-color;
+
+        #wrapper {
+            width: 100%;
+            height: 100vh;
+            @include flex(flex, column nowrap, center, center);
+
+            #content {
+                text-align: center;
+
+                h1 {
+                    font-size: 5rem;
+                    margin: 20px 0 20px 0;
+                    font-weight: bold;
+                }
+
+                h2 {
+                    font-weight: 700;
+                    font-size: 2rem;
+                    margin: 30px 0;
+                    color: white;
+                }
+
+                #socialmedia {
+                    margin: 30px auto;
+
+                    .fab {
+                        font-size: 1.8rem;
+                        margin: 0 10px;
+                        @include links($four-links-icon-f, $four-links-icon-f, $four-links-icon-fh, $four-links-icon-fh);
+                    }
+                }
+            }
+        }
+    }
+}
