+++ sass/layout/_home-main.scss
@@ -124,7 +124,7 @@ body {
           padding: 0 50px 0 0;
           text-align: right;
           line-height: 1.7;
-          font-size: 3.75rem;
+          font-size: 3.4rem;
           font-weight: bold;
           color: white;
         }
@@ -158,7 +158,7 @@ body {
           padding: 0 50px 0 0;
           text-align: right;
           line-height: 1.7;
-          font-size: 3.75rem;
+          font-size: 3.4rem;
           font-weight: bold;
           color: white;
         }
